Separate the authorication for api key load
