+++ titanium_white/__init__.py
@@ -1,3 +1,3 @@
-from .courtsession import CourtSession
+from .court_session import CourtSession
 
 __all__ = ["CourtSession"]
