+++ titanium_white/court_session.py
@@ -2,10 +2,10 @@
 Collection of classes for implementating the wrapper around
 the courtlistener API
 """
-import os
-
 from loguru import logger
 import requests
+
+from .authorization import load_api_key
 
 
 class CourtSession:
@@ -21,35 +21,9 @@ class CourtSession:
 
     def __init__(self, api_key: str = None):
         if api_key is None:
-            api_key = self.__load_api_key("COURTLISTENER")
+            api_key = load_api_key("COURTLISTENER")
         self.api_key = api_key
         self.http_session = requests.Session()
-
-    @classmethod
-    def __load_api_key(cls, courtlistener_env: str = None) -> str:
-        """
-        Attempts to load the api key from the environment variable
-        specified as the "courtlistener_env" argument. If no argument
-        is provided then we default to searching for the environment
-        variable "COURTLISTENER"
-        """
-        if courtlistener_env is None:
-            courtlistener_env = "COURTLISTENER"
-
-        try:
-            api_key = os.environ[courtlistener_env]
-        except KeyError as key_err:
-            logger.exception(key_err)
-            env_error_msg = (
-                "Missing COURTLISTENER API token\n"
-                "Either set the environment variable or pass"
-                "to the <api_key_instance> constructor argument"
-            )
-            logger.error(env_error_msg)
-            raise key_err
-        else:
-            logger.debug((f"Discovered api key " f"[{courtlistener_env}]={api_key}"))
-        return api_key
 
     def __enter__(self):
         """
